Moved the object helpers to the shared helpers
